requirements.txt
altair==5.5.0
google-generativeai>=0.10.0
attrs==25.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.41
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
multitasking==0.0.11
narwhals==1.30.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
protobuf==5.29.3
pyarrow==19.0.1
pydeck==0.9.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.43.2
streamlit-option-menu==0.4.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
yfinance==0.2.54
